Función formatDate a validatorsService
