+++ src/Services/validatorsService.js
@@ -1,0 +1,9 @@
+export function formatDate(dateString) {
+    const date = new Date(dateString);
+    const day = date.getDate();
+    const month = date.getMonth();
+    const year = date.getFullYear();
+    const hours = date.getHours();
+    const minutes = date.getMinutes();
+    return (`${day}-${month}-${year} ${hours}:${minutes}`)
+}
